backend/requirements.txt
starlette==0.37.2
uvicorn==0.32.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.17
SQLAlchemy[asyncio]
psycopg[binary]==3.3.3
alembic
argon2-cffi==23.1.0
